requirements.txt
altgraph==0.17.4
auto-py-to-exe==2.43.3
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2024.2.2
charset-normalizer==3.3.2
Eel==0.16.0
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
idna==3.6
macholib==1.16.3
packaging==24.0
Pillow==10.1.0
pycups==2.0.1
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
pyparsing==3.1.2
pypng==0.20220715.0
qrcode==7.4.2
requests==2.31.0
setuptools==69.2.0
tk==0.1.0
typing_extensions==4.10.0
urllib3==2.2.1
whichcraft==0.6.1
zope.event==5.0
zope.interface==6.2
